requirements.txt
numpy>=1.24.0,<2.0.0
opencv-python-headless>=4.8.0
tensorflow>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
streamlit>=1.27.0
pillow>=10.0.0
gdown>=4.7.0